Move AnalyzerScope
